hopefully fixed block incorrect relic place not working (needs actual testing) - hopefully fixed detecting other peoples placed relic time - fixed kickedTimer being triggered if someone sent the msg in chat (needs testing)
- added class colors for melody and location msgs (needs testing)
- added chat notifications (fix § not being able to be inputed in configString)
- added 1.8.9 rag sound toggle
- added mute bonzo staffs ghast sound

- did some refactoring

+++ src/main/java/blade/addon/utils/dungeon/Phase.java
@@ -161,10 +161,9 @@ public class Phase {
         Events.ON_RUN_END.invoke(RunEndEvent::onRunEnd);
     }
 
-    public static double getPhase() {
+    public static int getPhase() {
         return currentPhase;
     }
-
 
     public static boolean isInFloor7() {
         return inFloor7;
